Make width flexible above 960px
Width of content was previously fixed at 960px. 960px is now the
minimum width, but if there is space the content will expand to
fill 90% of the browser window.

Note that the entry-meta position has also changed to be properly
floated, as this is easier to manage and fits in better.

Change-Id: I1b9d01a832d5191768edc1764cded4a4b57fe2b5