Update patch set 6
Patch Set 6:

Many thanks for the review Jon, most helpful - I didn't realise the viewport tag was required for mobiles to take your design seriously ;)

"It would be great if ResourceLoader could create link tags with media attributes to only load css conditionally but currently it doesn't do that."

I imagine that would be tricky to implement well, as you'd want to load all the CSS for different media which *could* be activated, e.g. in a computer browser which can change size. I think it would be too much work (and too fragile) to be worth it. Not that I'd do that in this commit anyway :p

Patch-set: 6