Update patch set 1
Patch Set 1: Code-Review+2

I'm just going to go ahead and merge this.

I'm a little fuzzy as to what's going on with the navigation/menus in general with this skin, but this does seem to be an unambiguous improvement.

Patch-set: 1
Reviewer: Isarra <463@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2