Restore loading legacy styles removed in MediaWiki 1.35
Before MediaWiki 1.35, legacy styles were silently loaded for all skins.

These styles are 3.75kb in size and contain various CSS rules that skins
might have their own styles for already. It also contains CSS required
for localised (non-English) content, including for right-to-left
languages.

This is a good opportunity to check in if the styles are actually needed.
This patch shows how to keep the skin the same as before, and can be
reverted if they are not be needed by your skin.

More information on the change is given in ResourceLoaderSkinModule and
T242177.

Change-Id: I94fd2334b258472c96e3740f39a643b841ce3b6e